feat: crud of houses

+++ src/main/java/com/ifrn/alugo/service/HouseService.java
@@ -1,13 +1,11 @@
-package com.ifrn.domusmanager.service;
+package com.ifrn.alugo.service;
 
-import com.ifrn.domusmanager.dto.HouseRequestDTO;
-import com.ifrn.domusmanager.dto.HouseResponseDTO;
-import com.ifrn.domusmanager.entity.Address;
-import com.ifrn.domusmanager.entity.House;
-import com.ifrn.domusmanager.exceptions.ResourceNotFoundException;
-import com.ifrn.domusmanager.mappers.HouseMapper;
-import com.ifrn.domusmanager.repository.HouseRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.ifrn.alugo.dto.HouseRequestDTO;
+import com.ifrn.alugo.dto.HouseResponseDTO;
+import com.ifrn.alugo.entity.House;
+import com.ifrn.alugo.exceptions.ResourceNotFoundException;
+import com.ifrn.alugo.mappers.HouseMapper;
+import com.ifrn.alugo.repository.HouseRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,7 +13,6 @@ import java.util.stream.Collectors;
 
 @Service
 public class HouseService {
-
     private final HouseMapper houseMapper;
     private final HouseRepository houseRepository;
 
@@ -25,7 +22,7 @@ public class HouseService {
     }
 
     public HouseResponseDTO createHouse(HouseRequestDTO houseRequestDTO) {
-        House house = houseMapper.toEntityFromRequest(houseRequestDTO);
+        House house = houseMapper.toEntity(houseRequestDTO);
         houseRepository.save(house);
         return houseMapper.toResponseDTO(house);
     }
@@ -47,13 +44,14 @@ public class HouseService {
         House house = houseRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("House with id " + id + " not found"));
 
-        houseMapper.updateEntityFromRequest(houseRequestDTO, house);
-        return houseMapper.toResponseDTO(houseRepository.save(house));
+        house = houseMapper.updateEntityFromRequest(houseRequestDTO, house);
+        return houseMapper.toResponseDTO(houseRepository.saveAndFlush(house));
     }
 
     public void deleteHouse(Long id) {
-        House house = houseRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("House with id " + id + " not found"));
+        if (!houseRepository.existsById(id)) {
+            throw new ResourceNotFoundException("House with id " + id + " not found");
+        }
         houseRepository.deleteById(id);
     }
 }
